% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BART_call.R
\name{BART_call}
\alias{BART_call}
\title{Bayesian Additive Regression Trees}
\usage{
BART_call(formula, data, test.data = NULL, Prior = NULL, Mcmc = NULL)
}
\arguments{
\item{formula}{response ~ covariates,}

\item{test.data}{Test Data, typically without the response,}

\item{Prior}{List of Priors for MPBART: e.g., Prior = list(nu=3,sigq=0.9, ntrees=200,  kfac=2.0, pswap=0,  pbd=1.0, pb=0.5 , beta = 2.0, alpha = 0.95, nc = 100, priorindep = FALSE,  minobsnode = 10).
The components of Prior are
\itemize{
\item nu: The degree of freedom in the inverse chi-square prior distribution of the error variance.
\item sigq: The quantile of the error variance prior that the rough estimate (from linear regression) is placed at.
\item ntrees : The total number of trees in each round of BART fitting.
\item kfac : A tuning parameter that satisfies mu - kfac * sigma = ymin and mu + kfac * sigma = ymax, where mu and sigma are the mean and std of the Gaussian prior distribution of the sum of fit of all trees.
\item pswap : The prior probability of swap move in simulating trees; default 0, there should be pswap no larger than 1-pbd.
\item pbd : The prior probability of birth/death move in simulating trees; default 1.
\item pb : The prior probability of birth given birth/death; default 0.5.
\item alpha : The prior probability of a bottom node splits is alpha/(1+d)^beta, d is depth of node.
\item beta : see alpha.
\item nc : The number of equally spaced cutpoints between min and max for each covariate.
\item priorindep
\item minobsnode : The minimum number of observations in bottom nodes for birth in simulating trees.
}}

\item{Mcmc}{List of MCMC starting values, burn-in ...: e.g.,     list(burn = 100, ndraws = 1000)}

\item{train.data}{Training Data with the response,}
}
\value{
treefit_train ndraws x n posterior matrix of the training data sum of trees fit,

treefit_test ndraws x testn posterior matrix of the test data sum of trees fit,

samp_train ndraws x n posterior matrix of the training data outcome,

samp_test ndraws x testn posterior matrix of the test data outcome,

sigmasample posterior samples of the error standard deviation.
}
\description{
Bayesian Additive Regression Trees Modeling for Continuous Outcome,
}
