% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynBART_call.R
\name{DynBART_call}
\alias{DynBART_call}
\title{Bayesian Additive Regression Trees}
\usage{
DynBART_call(formula, data, test.data = NULL, Prior = NULL, Mcmc = NULL)
}
\arguments{
\item{formula}{response ~ covariates,}

\item{data}{Training data with the continuous response,}

\item{test.data}{Test data with number of rows equals nsamp x ndraws, typically without the response. For the ith subject, (i, i+nsamp, ..., i+(ndraws-1)nsamp) rows are its simulated posterior outcomes from the previous simulation,}

\item{Prior}{List of Priors for MPBART: e.g., Prior = list(nu=3,sigq=0.9, ntrees=200,  kfac=2.0, pswap=0,  pbd=1.0, pb=0.5 , beta = 2.0, alpha = 0.95, nc = 100, minobsnode = 10).
The components of Prior are
\itemize{
\item nu: The degree of freedom in the inverse chi-square prior distribution of the error variance.
\item sigq: The quantile of the error variance prior that the rough estimate (from linear regression) is placed at.
\item ntrees : The total number of trees in each round of BART fitting.
\item kfac : A tuning parameter that satisfies mu - kfac * sigma = ymin and mu + kfac * sigma = ymax, where mu and sigma are the mean and std of the Gaussian prior distribution of the sum of fit of all trees.
\item pswap : The prior probability of swap move in simulating trees; default 0, there should be pswap no larger than 1-pbd.
\item pbd : The prior probability of birth/death move in simulating trees; default 1.
\item pb : The prior probability of birth given birth/death; default 0.5.
\item alpha : The prior probability of a bottom node splits is alpha/(1+d)^beta, d is depth of node.
\item beta : see alpha.
\item nc : The number of equally spaced cutpoints between min and max for each covariate.
\item minobsnode : The minimum number of observations in bottom nodes for birth in simulating trees.
}}

\item{Mcmc}{List of MCMC starting values, burn-in ...: e.g.,     list(burn = 100, ndraws = 1000)}
}
\value{
treefit_train ndraws x n posterior matrix of the training data sum of trees fit,

treefit_test ndraws x testn posterior matrix of the test data sum of trees fit,

samp_train ndraws x n posterior matrix of the training data outcome,

samp_test ndraws x testn posterior matrix of the test data outcome,

sigmasample posterior samples of the error standard deviation.
}
\description{
Bayesian Additive Regression Trees Modeling for Continuous Outcome for Dynamic Data,
}
\examples{
##simulate data (example from Friedman MARS paper)
f = function(x){
 10*sin(pi*x[,1]*x[,2]) + 20*(x[,3]-.5)^2+10*x[,4]+5*x[,5]
}
sigma = 1.0 #y = f(x) + sigma*z , z~N(0,1)
n = 100 #number of observations
set.seed(99)
###############################################
nd = 1000 # ndraws
x1 = matrix(runif(n*9),n,9) #10 variables, only first 5 matter
x1 = x1[rep(1:n,nd),]
x2 = runif(n*nd) # simulated dynamic variable
x = cbind(x1, x2)
Ey = f(x)
y=Ey+sigma*rnorm(n*nd)
tdat = data.frame(x,y)
###############################################
x=matrix(runif(n*10),n,10) #10 variables, only first 5 matter
Ey = f(x)
y=Ey+sigma*rnorm(n)
dat = data.frame(x,y)
colnames(tdat) = colnames(dat)
###############################################
fml = as.formula("y ~ X1+X2+X3+X4+X5+X6+X7+X8+X9+X10")
bfit = DynBART_call(fml, data = dat, test.data = tdat,
                Prior = list(nu = 3, sigq = 0.9,
                             ntrees = 100,
                             kfac = 2,
                             pswap = 0.1, pbd = 0.5, pb = 0.25,
                             alpha = 0.95, beta = 2.0,
                             nc = 100, minobsnode = 10),
                Mcmc = list(burn=100, ndraws = nd))

}
