% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBART_call.R
\name{PBART_call}
\alias{PBART_call}
\title{Probit Bayesian Additive Regression Trees}
\usage{
PBART_call(formula, data, test.data = NULL, Prior = NULL, Mcmc = NULL)
}
\arguments{
\item{formula}{response ~ covariates,}

\item{test.data}{Test Data, typically without the response,}

\item{Prior}{List of Priors for MPBART: e.g., Prior = list(ntrees=200,  kfac=2.0, pswap=0,  pbd=1.0, pb=0.5 , beta = 2.0, alpha = 0.95, nc = 100, priorindep = FALSE,  minobsnode = 10).
The components of Prior are
\itemize{
\item ntrees : The total number of trees in each round of BART fitting.
\item kfac : A tuning parameter that satisfies mu - kfac * sigma = ymin and mu + kfac * sigma = ymax, where mu and sigma are the mean and std of the Gaussian prior distribution of the sum of fit of all trees.
\item pswap : The prior probability of swap move in simulating trees; default 0, there should be pswap no larger than 1-pbd.
\item pbd : The prior probability of birth/death move in simulating trees; default 1.
\item pb : The prior probability of birth given birth/death; default 0.5.
\item alpha : The prior probability of a bottom node splits is alpha/(1+d)^beta, d is depth of node.
\item beta : see alpha.
\item nc : The number of equally spaced cutpoints between min and max for each covariate.
\item priorindep
\item minobsnode : The minimum number of observations in bottom nodes for birth in simulating trees.
}}

\item{Mcmc}{List of MCMC starting values, burn-in ...: e.g.,     list(burn = 100, ndraws = 1000)}

\item{train.data}{Training Data with the response taking values 0-1,}
}
\value{
treefit_train ndraws x n posterior matrix of the training data sum of trees fit,

treefit_test ndraws x testn posterior matrix of the test data sum of trees fit,

samp_train ndraws x n posterior matrix of the training data outcome,

samp_test ndraws x testn posterior matrix of the test data outcome.
}
\description{
Bayesian Additive Regression Trees Modeling for Binary Outcome,
}
